+++ packages/reakit/src/Card/index.ts
@@ -1,14 +1,9 @@
 import Card from "./Card";
 import CardFit from "./CardFit";
 
-interface CardComponents {
-  Fit: typeof CardFit;
-}
-
-const C = Card as typeof Card & CardComponents;
-
-C.Fit = CardFit;
-
 export * from "./Card";
+export * from "./CardFit";
 
-export default C;
+export default Object.assign(Card, {
+  Fit: CardFit
+});