+++ packages/reakit/src/Group/Group.ts
@@ -1,4 +1,4 @@
-import PropTypes from "prop-types";
+import * as PropTypes from "prop-types";
 import { ifProp, prop, theme, withProp } from "styled-tools";
 import styled, { css } from "../styled";
 import getSelector from "../_utils/getSelector";
@@ -8,7 +8,7 @@ import GroupItem from "./GroupItem";
 
 const groupItemSelector = getSelector(GroupItem);
 
-const verticalAt = (pass, fail) =>
+const verticalAt = (pass: any, fail?: any) =>
   ifProp(
     "verticalAt",
     css`
@@ -22,7 +22,12 @@ const verticalAt = (pass, fail) =>
     fail
   );
 
-const Group = styled(Box)`
+export interface GroupProps {
+  vertical: boolean;
+  verticalAt: number;
+}
+
+const Group = styled(Box)<GroupProps>`
   display: flex;
   flex-direction: ${ifProp("vertical", "column", "row")};
   ${verticalAt("flex-direction: column")};
@@ -69,6 +74,7 @@ const Group = styled(Box)`
   ${theme("Group")};
 `;
 
+// @ts-ignore
 Group.propTypes = {
   vertical: PropTypes.bool,
   verticalAt: PropTypes.number