+++ packages/reakit/src/Step/__tests__/Step-test.tsx
@@ -1,20 +1,21 @@
-import React from "react";
+import * as React from "react";
 import { mount } from "enzyme";
-import Step from "../Step";
+import Step, { StepProps } from "../Step";
 import Hidden from "../../Hidden";
 
-const props = {
+const props: StepProps = {
   step: "foo",
   order: 0,
   register: jest.fn(),
   update: jest.fn(),
-  unregister: jest.fn()
+  unregister: jest.fn(),
+  isCurrent: jest.fn().mockReturnValue(true)
 };
 
 beforeEach(() => {
-  Object.keys(props).forEach(key => {
-    if (props[key].mockClear) {
-      props[key].mockClear();
+  Object.values(props).forEach(value => {
+    if (value.mockClear) {
+      value.mockClear();
     }
   });
 });
@@ -54,6 +55,6 @@ it("has truthy visible property when it is current", () => {
 });
 
 it("has falsy visible property when it is not current", () => {
-  const wrapper = mount(<Step {...props} />);
+  const wrapper = mount(<Step {...props} isCurrent={() => false} />);
   expect(wrapper.find(Hidden).prop("visible")).toBe(false);
 });