+++ packages/reakit/test.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable no-unused-expressions */
 import * as React from "react";
-import { styled, Box, Code, Card } from "./src";
+import { styled, Box, Code, Card, Step } from "./src";
 
 interface Props {
   foo: string;
@@ -17,4 +17,13 @@ interface Props {
   <Code />;
   <Code block />;
   <Card gutter={8} />;
+
+  <Step.Container>
+    {step => (
+      <div>
+        <Step.Hide hide={step.hide} />
+        <Step step="test" {...step} />
+      </div>
+    )}
+  </Step.Container>;
 }